Guard AOF page-flush test hook behind #if DEBUG; split injected test into its own DEBUG-only file
Copilot-Session: a8557ca8-3a4c-487c-bbf1-dd45e03caca5
